% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{convolution}
\alias{convolution}
\title{Add discrete stochastic random variables of arbitrary dimensions}
\usage{
convolution(dat, threshold = NA, verbose = F)
}
\arguments{
\item{dat}{A data.table with columns 'x', 'y', and 'probabilities'. Column 'x' is a numeric indicator unique for each random variable to be convoluted together. Column 'y' is the outcome value, and colum 'probabilities' contains the probability corresponding to each value of the 'y' column.}

\item{threshold}{An optional number to set a threshold. If set, the probabilities of all possible outcomes of the sum of the random variables are only caluclated below this value.}

\item{verbose}{A logical indicator which defaults to FALSE. Set to TRUE to print information.}
}
\value{
the convoluted distribution.
}
\description{
Add discrete stochastic random variables of arbitrary dimensions
}
\examples{
convolution(dat = data.table(x = c(rep(1, 2), rep(2, 3), rep(3, 4)), y = c(1:2, 1:3, 1:4), 
    probability = c(3/4, 1/4, rep(1/3, 3), (1:4)/10)))
convolution(dat = data.table(x = c(rep(1, 2), rep(2, 3), rep(3, 4)), y = c(1:2, 1:3, 1:4), 
    probability = c(3/4, 1/4, rep(1/3, 3), (1:4)/10)), threshold = 6)
}

