% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdDetect.R
\name{ncdDetect}
\alias{ncdDetect}
\title{Add discrete stochastic random variables of same dimension}
\usage{
ncdDetect(predictions, scores, observations = NA, thres = NA)
}
\arguments{
\item{predictions}{A matrix in which each row corresponds to a discrete random variable and each column corresponds to a discrete outcome. The matrix contains the probabilities of each outcome, and each row must sum to one.}

\item{scores}{A matrix in which each row corresponds to a discrete random variable and each column corresponds to a discrete outcome. The matrix contains the values for each outcome for each random variable.}

\item{observations}{A matrix (optional) in which each row corresponds to a discrete random variable and each column corresponds to a discrete outcome. The matrix contains the observed outcome for each variable. Each row must contain exactly one 1, while the rest of the entries must be 0.}

\item{thres}{An optional number to set a threshold. If set, the probabilities of all possible outcomes of the sum of the random variables are only caluclated below this value.}
}
\value{
score_dist The convoluted distribution of the sum of the discrete random variables defined by input matrices predictions and scores.

obs_score The observed score. Only returned if matrix observations is provided as input.

p_value The p-value resulting from evaluating the observed score in the convoluted distribution. Only returned if matrix observations is provided as input.
}
\description{
Add discrete stochastic random variables of same dimension
}
\examples{
ncdDetect(predictions = matrix(rep(1/6, 12), nrow = 2), scores = matrix(rep(1:6, 2), nrow = 2, byrow = TRUE), 
    observations = matrix(c(0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0), nrow = 2, byrow = TRUE))
ncdDetect(predictions = matrix(rep(1/6, 12), nrow = 2), scores = matrix(rep(1:6, 2), nrow = 2, byrow = TRUE), thres = 6)
ncdDetect(predictions = matrix(rep(1/6, 12), nrow = 2), scores = matrix(rep(1:6, 2), nrow = 2, byrow = TRUE))
}

